refactor: Move useDarkMode into global provider
