+++ src/components/heros/HeroConnectWalletButton.vue
@@ -1,5 +1,5 @@
 <script lang="ts" setup>
-import useDarkMode from '@/composables/useDarkMode';
+import { useDarkMode } from '@/providers/dark-mode.provider';
 import useFathom from '@/composables/useFathom';
 import useWeb3 from '@/services/web3/useWeb3';
