+++ src/composables/usePoolHelpers.ts
@@ -1,6 +1,5 @@
 import { AprBreakdown, PoolType } from '@balancer-labs/sdk';
 import { getAddress } from '@ethersproject/address';
-
 import { APR_THRESHOLD } from '@/constants/pools';
 import { Network } from '@/lib/config/types';
