+++ src/services/web3/connectors/trustwallet/walletconnect.connector.ts
@@ -4,7 +4,7 @@ import { configService } from '@/services/config/config.service';
 import { WalletError } from '@/types';
 import { Connector, ConnectorId } from '../connector';
 import { Network } from '@/lib/config/types';
-import useDarkMode from '@/composables/useDarkMode';
+import { useDarkMode } from '@/providers/dark-mode.provider';
 
 const { MAINNET, ARBITRUM, AVALANCHE, BASE, GNOSIS, POLYGON, ZKEVM } = Network;
